requirements.txt
ipykernel
ipython_pygments_lexers
matplotlib
monai
opencv-python
openpyxl
pandas
pip
PyDrive
python-dotenv
scikit-learn
timm
torchmetrics
wandb
wheel
